Added nodemon dependency
