+++ src/package.json
@@ -2,7 +2,7 @@
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
-    "dev": "node server/main.js",
+    "dev": "nodemon server/main.js",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
@@ -18,6 +18,7 @@
     "expo-status-bar": "~1.0.4",
     "express": "^4.17.1",
     "graphql": "^15.5.0",
+    "nodemon": "^2.0.7",
     "pg": "^8.6.0",
     "react": "16.13.1",
     "react-dom": "16.13.1",
