+++ src/server/models/dummyData.js
@@ -1,0 +1,25 @@
+const userData = {
+  1: {
+    id: '1',
+    name: 'testUser',
+    email: 'test@gmail.com',
+  },
+  2: {
+    id: '2',
+    name: 'testUser2',
+    email: 'test2@gmail.com',
+  },
+};
+
+const projectsData = {
+  1: {
+    id: '1',
+    name: 'GitStop',
+  },
+  2: {
+    id: '2',
+    name: 'testProj',
+  },
+};
+
+module.exports = { userData, projectsData };
