+++ src/server/resolvers/resolvers.js
@@ -1,16 +1,15 @@
-const { ApolloServer } = require('apollo-server-express');
-const { schema, userData, projectsData } = require('../schema/databaseSchema.js')
+// const { ApolloServer } = require('apollo-server-express');
+// const { schema, userData, projectsData } = require('../schema/databaseSchema.js')
 
-const resolvers = {
-  Query: {
-    users: () => {
-      return Object.values(userData);
-      },
-    user: () => {
-      return name;
-    }
-  },
-};
+// const resolvers = {
+//   Query: {
+//     users: () => {
+//       return Object.values(userData);
+//       },
+//     user: () => {
+//       return name;
+//     }
+//   },
+// };
 
-module.exports = resolvers;
-
+// module.exports = resolvers;
